A remake of the classic Arkanoid, written in under a month by 4 university students.
Project configured for CodeBlocks 16.01. It uses SDL2 along a few extension libraries, SDL_Mixer for audio and SDL_Image for graphics. All used assets (sounds, sprites, and other resources) are properties of their rightful owners. If facing any compilation issues, make sure the standard is set to at least C++11. The videogame was created as part of the final project for a course on Object-Oriented Programming, Fall '17 Habib University. I acknowledge the fact that there are a few loopholes and inefficiencies in the code, but just putting this up here for everyone to see/contribute.

A game by:
Osama Yousuf
Amin Shiraz Gillani
Syed Sameer Nadeem
Minhaj Ahmed Moin